% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-estimation.R
\name{add_qlearner}
\alias{add_qlearner}
\title{Add a Q-learner to a list of learners}
\usage{
add_qlearner(
  learners,
  name,
  type = c("ql", "drql", "ptl", "policytree"),
  action_name = "A",
  covariates = c("x1", "x2"),
  q_func = NULL,
  sl_library = NULL,
  depth = 2,
  hybrid = FALSE
)
}
\arguments{
\item{learners}{List to which the learner configuration will be added.}

\item{name}{String indicating the unique name of the expert.}

\item{type}{String indicating the type of learner.
("ql", "drql", "ptl", "policytree").}

\item{action_name}{String indicating the treatment variable name. Defaults to "A".}

\item{covariates}{Character vector of covariate names. Defaults to \code{c("x1", "x2")}.}

\item{q_func}{String indicating the type of function for Q-learning.
("q_glm", "q_rf", "q_xgboost", "q_sl"). Defaults to NULL.}

\item{sl_library}{Vector of libraries for SuperLearner.
Required if \code{q_func} is "q_sl". Defaults to NULL.}

\item{depth}{Numeric value for the depth of growing trees. Defaults to 2.}

\item{hybrid}{Logical indicating whether to use a hybrid approach. Defaults to FALSE.}
}
\value{
The updated \code{learners} list with the new learner and its characteristics.
}
\description{
Appends a learner configuration to a list and automatically generates the
associated \code{q_model} and \code{qv_model} using \code{polle} specifications.
}
\examples{
\dontrun{
learners <- list()
learners <- add_qlearner(
  learners, 
  name = "ql_lm", 
  type = "ql", 
  q_func = "q_glm", 
  action_name = "A"
)
}
}
