% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{baseline_effect}
\alias{baseline_effect}
\title{Baseline Outcome Effect Function}
\usage{
baseline_effect(X)
}
\arguments{
\item{X}{A numeric matrix of covariates of size n x 2.}
}
\value{
A numeric vector of length n containing the baseline effects.
}
\description{
Calculates the non-treatment related baseline component of the outcome model
based on a linear combination and an exponential transformation of covariates.
}
\examples{
X <- matrix(stats::runif(10 * 2), 10, 2)
baseline_effect(X)
}
