% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{coverage_strict}
\alias{coverage_strict}
\title{Strict coverage}
\usage{
coverage_strict(true_set, pred_set, levels = 1:5)
}
\arguments{
\item{true_set}{A \code{list} of numeric or character vectors representing the ground truth sets.}

\item{pred_set}{A \code{list} of numeric or character vectors representing the predicted sets.}
}
\value{
A numeric value representing the mean coverage (proportion of
intersected elements over true set size) across all observations.
}
\description{
Computes the average proportion of the true set that is strictly contained within
the predicted set. This is a normalized measure of recall across multiple
observations.
}
\examples{
true <- list(c(1, 2), c(3))
pred <- list(c(1), c(3, 4))
coverage_strict(true, pred)
}
