% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{coverage_strict_single}
\alias{coverage_strict_single}
\title{Strict coverage for single observation}
\usage{
coverage_strict_single(true_set, pred_set)
}
\arguments{
\item{true_set}{A \code{list} of numeric or character vectors representing the ground truth sets.}

\item{pred_set}{A \code{list} of numeric or character vectors representing the predicted sets.}
}
\value{
A numeric value representing the mean coverage (proportion of
intersected elements over true set size) across all observations.
}
\description{
Returns one if true set is strictly contained within
the predicted set zero otherwose.
}
\examples{
true <- list(c(1, 2), c(3))
pred <- list(c(1), c(3, 4))
coverage_strict(true, pred)
}
