% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_data}
\alias{generate_data}
\title{Synthetic data generator}
\usage{
generate_data(
  n,
  ncov = 2,
  seed = NA,
  type = c("normal", "complex"),
  is_RCT = TRUE
)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{ncov}{Number of covariates (2 by default).}

\item{seed}{Integer or NA (NA by default).}

\item{type}{String indicating the type of synthetic scenario ("normal", "complex")}

\item{is_RCT}{Logical indicating if treatment allocated as in RCT (TRUE by default).}
}
\value{
A list containing two data frames (\code{df_obs} with observed outcomes
based on treatment and \code{df_complete} with all potential outcomes) and the
oracular optimal treatment assignments.
}
\description{
Generates a dataset simulating treatment assignment, covariates, and potential outcomes.
}
\examples{
n <- 1e3 
generate_data(n, type="normal")
}
