% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-estimation.R
\name{glearner}
\alias{glearner}
\title{G-model creator}
\usage{
glearner(m, g_func = NULL, sl_library = NULL, num.trees = NULL)
}
\arguments{
\item{m}{Numeric value for the number of treatment levels.}

\item{g_func}{String indicating the type of function for G-learning.
("g_glm", "g_glmnet", "g_rf", "g_xgboost", and "q_sl").}

\item{sl_library}{String vector of libraries for learning policies
using SuperLearner. Required if g_func = "g_sl".}

\item{num.trees}{Number of trees for growing trees.  Defaults to NULL.
Required if g_func = "g_rf". Defaults to NULL.}
}
\value{
Creates a learner
}
\description{
Generates the associated \code{g_model} using \code{polle} specifications.
}
\examples{
\dontrun{
g_model <- glearner(
  m = 5, 
  g_func = "g_glm")
}
}
