% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{heatmap_treatments}
\alias{heatmap_treatments}
\title{Create a heatmap for treatments included in set-valued policies}
\usage{
heatmap_treatments(confidence_set, levels_A)
}
\arguments{
\item{confidence_set}{Integer index for the mechanism.}

\item{levels_A}{String label for the mechanism.}
}
\description{
Creates a data frame one-hot encoding the inclusion of treatments
in set-valued policies. Rows represent observations and columns
the treatment levels.
}
