% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{ivf_set_policy_values}
\alias{ivf_set_policy_values}
\title{Set-policy values for IVF data example}
\usage{
ivf_set_policy_values(
  test_set,
  test,
  covariates = c("x1", "x2"),
  treatment_name = "A",
  outcome_name = "Y",
  second_outcome = "xi",
  mod_y,
  mod_xi,
  mod_ps,
  ab,
  ab_xi,
  n_test = 1,
  levels
)
}
\arguments{
\item{test_set}{A \code{list} of numeric or character vectors representing the ground truth sets.}

\item{test}{Data frame used for prediction (calibration/test set).}

\item{covariates}{Character vector of covariate names. Defaults to \code{c("x1", "x2")}.}

\item{treatment_name}{String indicating the treatment variable. Defaults to "A".}

\item{outcome_name}{String indicating the outcome variable. Defaults to "Y".}

\item{second_outcome}{String indicating the second outcome variable. Defaults to "xi".}

\item{mod_y}{Model to predict the conditional mean of the second outcome.}

\item{mod_ps}{Model to predict the propensity score.}

\item{ab}{Float indicating the largest difference in the outcome.}

\item{ab_xi}{Float indicating the largest difference in the second outcome.}

\item{n_test}{Integer indicating the number of policies to sample at random for the set-valued policy.}

\item{levels}{Vector of possible treatment/action levels. Defaults to \code{1:5}.}
}
\value{
A list with the estimated set-policy values (random and lowest
strategy for Y and xi) of the set-valued policy.
}
\description{
Estimates the uniform set-policy value for a primary outcome (Y) and an
adverse event (xi). Additionally computes the value for a "minimal treatment"
strategy—selecting the lowest available treatment level across all cases.
\strong{Note:} This function is uses the package \code{SL.ODTR} from
Montoya, L. M., van der Laan, M. J., Luedtke, A. R., Skeem, J. L., Coyle, J. R.,
& Petersen, M. L. (2023). The optimal dynamic treatment rule superlearner:
considerations, performance, and application to criminal justice interventions.
}
