% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_block}
\alias{make_block}
\title{Create a standard data block}
\usage{
make_block(mech_index, mech_name, array_a, alpha_vec, rate_data)
}
\arguments{
\item{mech_index}{Integer index for the mechanism.}

\item{mech_name}{String label for the mechanism.}

\item{array_a}{4D array with dimensions: (level, rep, mechanism, rate).}

\item{alpha_vec}{Vector of alpha levels.}

\item{rate_data}{Vector of rate labels.}
}
\description{
Reshapes a 4D array into a long-format data frame.
}
