% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-estimation.R
\name{make_g_model}
\alias{make_g_model}
\title{Builder of G-model for Policy Learning}
\usage{
make_g_model(g_func, sl_library = NULL, num.trees = NULL)
}
\arguments{
\item{g_func}{String indicating the type of function for Q-learning.
("g_glm", "g_glmnet", "g_rf", "g_xgboost", and "q_sl").}

\item{sl_library}{String vector of libraries for learning policies
using SuperLearner. Required if g_func = "g_sl". Defaults to NULL.}

\item{num.trees}{Number of trees for growing trees.
Required if g_func = "g_rf".  Defaults to NULL.}
}
\value{
An object of class q_model as defined by the polle package.
}
\description{
A wrapper function to build G-models required for training policies using the
\code{polle} package.
}
\examples{
\dontrun{
make_g_model(g_func = "g_glm")
}
}
