% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-estimation.R
\name{make_q_model}
\alias{make_q_model}
\title{Q-model builder for Policy Learning}
\usage{
make_q_model(
  q_func,
  covariates = c("x1", "x2"),
  sl_library = NULL,
  action_name = "A"
)
}
\arguments{
\item{q_func}{String indicating the type of function for Q-learning.
c("q_glm", "q_rf", "q_xgboost", and "q_sl").}

\item{covariates}{Character vector of covariate names. Defaults to \code{c("x1", "x2")}.}

\item{sl_library}{String vector of libraries for learning policies
using SuperLearner. Required if q_func = "q_sl".}

\item{action_name}{String indicating the treatment/action variable
name in the data frame. Defaults to "A".}
}
\value{
An object of class q_model as defined by the polle package.
}
\description{
A wrapper function to build Q-models required for training policies using the
polle package. It simplifies the creation of formula objects based on
treatment restrictions.
}
\examples{
\dontrun{
make_q_model(q_func = "q_glm", action_name = "treatment")
}
}
