% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{margin_score}
\alias{margin_score}
\title{Margin Nonconformity Score}
\usage{
margin_score(potential_outcomes)
}
\arguments{
\item{potential_outcomes}{Matrix of potential outcomes (observations by treatments).}
}
\value{
A matrix of the same dimensions as \code{potential_outcomes} containing
the margin nonconformity scores.
}
\description{
Generates the margin nonconformity scores for a matrix of potential outcomes.
The score is calculated as the difference between the maximum potential
outcome and all other outcomes, with a specific "margin" calculation for
the winning class.
}
\examples{
margin_score(matrix(runif(10 * 5), 10, 5))
}
