% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{mu_P0_normal}
\alias{mu_P0_normal}
\title{Conditional Mean Outcome: Linear Frontier Scenario}
\usage{
mu_P0_normal(X)
}
\arguments{
\item{X}{A numeric matrix of covariates of size n x 2.}
}
\value{
A numeric matrix of size n x 5, where each column represents the conditional mean for a specific treatment.
}
\description{
Simulates the conditional mean of the outcome under a "normal" scenario.
It creates a linear boundary that partitions which
treatment (1-2 vs. 3-4) receive an additive effect boost.
}
\examples{
X <- matrix(stats::runif(10 * 2), 10, 2)
mu_P0_normal(X)
}
