% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{oracular_set_policy_value}
\alias{oracular_set_policy_value}
\title{Oracular set-policy value}
\usage{
oracular_set_policy_value(
  test_set,
  test,
  test_potential_outcome,
  covariates = c("x1", "x2"),
  treatment_name = "A",
  outcome_name = "Y",
  n_test = 1,
  levels = 1:5
)
}
\arguments{
\item{test_set}{A \code{list} of numeric or character vectors representing the ground truth sets.}

\item{test}{Data frame used for prediction (calibration/test set).}

\item{test_potential_outcome}{A data frame containing the potential outcomes.}

\item{covariates}{Character vector of covariate names. Defaults to \code{c("x1", "x2")}.}

\item{treatment_name}{String indicating the treatment variable. Defaults to "A".}

\item{outcome_name}{String indicating the outcome variable. Defaults to "Y".}

\item{n_test}{Integer indicating the number of policies to sample at random for the set-valued policy.}

\item{levels}{Vector of possible treatment/action levels. Defaults to \code{1:5}.}
}
\value{
A numeric value representing the oracular set-policy value of the set-valued policy.
}
\description{
Computes the oracular set-policy value of a set-valued policy using potential outcomes.
\strong{Note:} This function is uses the package \code{SL.ODTR} from
Montoya, L. M., van der Laan, M. J., Luedtke, A. R., Skeem, J. L., Coyle, J. R.,
& Petersen, M. L. (2023). The optimal dynamic treatment rule superlearner:
considerations, performance, and application to criminal justice interventions.
}
