% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{set_policy_value}
\alias{set_policy_value}
\title{Set-policy value}
\usage{
set_policy_value(
  test_set,
  test,
  covariates = c("x1", "x2"),
  treatment_name = "A",
  outcome_name = "Y",
  mod_y,
  mod_ps,
  ab,
  n_test = 1,
  levels
)
}
\arguments{
\item{test_set}{A \code{list} of numeric or character vectors representing the ground truth sets.}

\item{test}{Data frame used for prediction (calibration/test set).}

\item{covariates}{Character vector of covariate names. Defaults to \code{c("x1", "x2")}.}

\item{treatment_name}{String indicating the treatment variable. Defaults to "A".}

\item{outcome_name}{String indicating the outcome variable. Defaults to "Y".}

\item{mod_y}{Model to predict the conditional mean outcome (estimate potential outcomes).}

\item{mod_ps}{Model to predict the propensity score.}

\item{ab}{Float indicating the largest difference in the outcome.}

\item{n_test}{Integer indicating the number of policies to sample at random for the set-valued policy.}

\item{levels}{Vector of possible treatment/action levels. Defaults to \code{1:5}.}
}
\value{
A numeric value representing the estimated set-policy value of the set-valued policy.
}
\description{
Estimates the uniform set-policy value of a set-valued policy.
\strong{Note:} This function is uses the package \code{SL.ODTR} from
Montoya, L. M., van der Laan, M. J., Luedtke, A. R., Skeem, J. L., Coyle, J. R.,
& Petersen, M. L. (2023). The optimal dynamic treatment rule superlearner:
considerations, performance, and application to criminal justice interventions.
}
