% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{width}
\alias{width}
\title{Average width of prediction set}
\usage{
width(pred_set, levels = 1:5)
}
\arguments{
\item{levels}{Vector of possible treatment levels. Defaults to \code{1:5}.}

\item{true_set}{A \code{list} of numeric or character vectors representing the ground truth sets.}
}
\value{
A numeric value representing the mean of the prediction set across all observations.
}
\description{
Computes the mean width the predicted set.
}
\examples{
pred <- list(c(1), c(3, 4))
width(pred)
}
